淡い感じ

+++ GLTelemetryPict/base/Renderers/YO/flTinyGLUtils/flTinyGLUtils/MultiColor/flMCGenerator.cpp
@@ -6,8 +6,6 @@
 //  Copyright (c) 2015 YoshitoONISHI. All rights reserved.
 //
 
-#include <stdio.h>
-#include <math.h>
 #include "flMCGenerator.h"
 #include "flMCNoise.h"
 
@@ -18,16 +16,16 @@ MULTI_COLOR_NAMESPACE_BEGIN
 
 void multicolor(Vector texture,
                 Color *color,
-                double arg0, double arg1, double arg2, double arg3,
-                double arg4, double arg5, double arg6, double arg7,
-                double arg8, double arg9, double arg10 )
+                float arg0, float arg1, float arg2, float arg3,
+                float arg4, float arg5, float arg6, float arg7,
+                float arg8, float arg9, float arg10 )
 {
     Vector axis, purt, cvec;
     Matrix matrix, tmpmat;
     
     copyMatrix(kIdentityMatrix, matrix);
     
-    axis = wrinkled(texture, 2.0, arg4, arg5);
+    axis = wrinkled(texture, 2.0f, arg4, arg5);
     
     rotateX(arg6 * axis.x, tmpmat);
     multMatrix(matrix, tmpmat, matrix);
@@ -37,15 +35,15 @@ void multicolor(Vector texture,
     multMatrix(matrix, tmpmat, matrix);
     
     purt = texture;
-    SMULT(0.3, purt);
-    purt  = wrinkled(purt, 2.0, 0.5, 7.0);
+    SMULT(0.3f, purt);
+    purt  = wrinkled(purt, 2.0f, 0.5f, 7.0f);
     SMULT(arg7, purt);
     VADD(texture, purt);
     
     cvec.x = arg10 * multifractal(texture, arg0, arg1, arg2, arg3);
-    texture.x += 10.5;
+    texture.x += 10.5f;
     cvec.y = arg10 * multifractal(texture, arg0, arg1, arg2, arg3);
-    texture.y += 10.5;
+    texture.y += 10.5f;
     cvec.z = arg10 * multifractal(texture, arg0, arg1, arg2, arg3);
     
     cvec = vectTransform(cvec, matrix);
@@ -54,27 +52,27 @@ void multicolor(Vector texture,
     color->g += arg8 * cvec.y;
     color->b += arg8 * cvec.z;
     
-    if (color->r < 0.0)
-        color->r = 0.0;
-    else if (color->r > 1.0)
-        color->r = 1.0;
-    if (color->g < 0.0)
-        color->g = 0.0;
-    else if (color->g > 1.0)
-        color->g = 1.0;
-    if (color->b < 0.0)
-        color->b = 0.0;
-    else if (color->b > 1.0)
-        color->b = 1.0;
+    if (color->r < 0.0f)
+        color->r = 0.0f;
+    else if (color->r > 1.0f)
+        color->r = 1.0f;
+    if (color->g < 0.0f)
+        color->g = 0.0f;
+    else if (color->g > 1.0f)
+        color->g = 1.0f;
+    if (color->b < 0.0f)
+        color->b = 0.0f;
+    else if (color->b > 1.0f)
+        color->b = 1.0f;
 }
 
-Vector wrinkled(Vector point, double lacunarity, double H, double octaves)
+Vector wrinkled(Vector point, float lacunarity, float H, float octaves)
 {
     Vector sPoint, result, temp;
-    double f, s;
+    float f, s;
     
-    result.x = result.y = result.z = 0.0;
-    f = s = 1.0;
+    result.x = result.y = result.z = 0.0f;
+    f = s = 1.0f;
     for (int i = 0; i < octaves; i++) {
         sPoint.x = f * point.x;
         sPoint.y = f * point.y;
@@ -89,14 +87,14 @@ Vector wrinkled(Vector point, double lacunarity, double H, double octaves)
     return result;
 }
 
-double multifractal(Vector pos,
-                    double H,
-                    double lacunarity,
-                    double octaves,
-                    double zero_offset)
+float multifractal(Vector pos,
+                    float H,
+                    float lacunarity,
+                    float octaves,
+                    float zero_offset)
 {
-    double y = 1.0;
-    double f = 1.0;
+    float y = 1.0f;
+    float f = 1.0f;
     
     for (int i = 0; i < octaves; i++) {
         y *= zero_offset + f * noise3(pos);
