+++ GLTelemetryPict/base/Renderers/YO/flTinyGLUtils/flTinyGLUtils/MultiColor/flMCTypes.h
@@ -20,26 +20,26 @@ namespace fl {
 MULTI_COLOR_NAMESPACE_BEGIN
 
 struct Color {
-    double r;
-    double g;
-    double b;
+    float r;
+    float g;
+    float b;
 };
 
 struct Vector {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
-typedef double Matrix[4][4];
+typedef float Matrix[4][4];
 
 extern const Matrix kIdentityMatrix;
 
 void copyMatrix(const Matrix a, Matrix b);
 void multMatrix(const Matrix a, const Matrix b, Matrix c);
-void rotateX(double theta, Matrix m);
-void rotateY(double theta, Matrix m);
-void rotateZ(double theta, Matrix m);
+void rotateX(float theta, Matrix m);
+void rotateY(float theta, Matrix m);
+void rotateZ(float theta, Matrix m);
 Vector vectTransform(const Vector v, const Matrix m);
 
 MULTI_COLOR_NAMESPACE_END
